Constructor of a model object

This function constructs an object of class fHMM_model, which contains details about the fitted (hierarchical) Hidden Markov model.

Usage

fHMM_model(
  data,
  estimate,
  nlm_output,
  estimation_time,
  ll,
  lls,
  gradient,
  inverse_fisher,
  decoding
)

# S3 method for class 'fHMM_model'
print(x, ...)

# S3 method for class 'fHMM_model'
residuals(object, ...)

# S3 method for class 'fHMM_model'
summary(object, alpha = 0.05, ...)

# S3 method for class 'fHMM_model'
coef(object, alpha = 0.05, digits = 2, ...)

# S3 method for class 'fHMM_model'
AIC(object, ..., k = 2)

# S3 method for class 'fHMM_model'
BIC(object, ...)

# S3 method for class 'fHMM_model'
nobs(object, ...)

# S3 method for class 'fHMM_model'
logLik(object, ...)

npar(object, ...)

# S3 method for class 'fHMM_model'
npar(object, ...)

# S3 method for class 'fHMM_model'
predict(object, ahead = 5, alpha = 0.05, ...)

Arguments

data

An object of class fHMM_data.

estimate

A numeric vector of unconstrained model estimates.

nlm_output

The output of nlm for the selected optimization run.

estimation_time

A diff.time object, the total estimation time.

ll

A numeric, the model log-likelihood.

lls

A numeric vector, the model log-likelihoods in all optimization runs.

gradient

A numeric vector, the gradient at the optimum.

inverse_fisher

A numeric vector, the inverse Fisher information for each parameter.

decoding

A numeric vector, the decoded time series.

x, object

An object of class fHMM_model.

...

Currently not used.

alpha

A numeric between 0 and 1, the confidence level.

digits

The number of decimal places.

k

Passed on to AIC.

ahead

The number of time points to predict ahead.

Value

An object of class fHMM_model.