Fit (hierarchical) hidden Markov models to financial data

Performs data processing, fitting, state decoding and visualization.

fit_hmm(controls, events, sim_par)

Arguments

controls
A list of controls (optional).

events
A list of (historical, financial) events (optional).

sim_par
A list of model parameters for simulation in thetaList format, default NULL (optional).

Value

No return value. Estimation results are saved in "controls[["path"]]/models/controls[["id"]]".

Details

Specify a model by setting parameters of the named list controls and passing it to fit_hmm. See the vignettes on how to specify controls.

Examples

### fitting a 2-state HMM with state-dependent t-distributions to simulated data
controls = list(
  path    = tempdir(),
  id      = "test",
  model   = "hmm",
  states  = 2,
  sdds    = "t",
  horizon = 200,
  fit     = list("runs" = 10, "seed" = 1)
)
fit_hmm(controls)

#> Controls checked.

#> - path: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpfWSJvd
#> - model id: test
#> - model type: hmm
#> - data type: simulated
#> - number of states: 2
#> - SDDs: t
#> - number of runs: 10 
#> - seed: 1

#> Data processed.

#> - sample size: 200

#> Selecting start values...

#> Error in .Call("_fHMM_LL_HMM_Rcpp", PACKAGE = "fHMM", allprobs, Gamma, : "_fHMM_LL_HMM_Rcpp" not available for .Call() for package "fHMM"