Prediction

This function predicts the next ahead states and data points based on an fHMM_model object.

Usage

# S3 method for fHMM_model
predict(object, ahead = 5, alpha = 0.05, ...)

Arguments

object

An object of class fHMM_model.

ahead

A positive integer, the forecast horizon.

alpha

The alpha level for the confidence interval, a numeric between 0 and 1. Per default, alpha = 0.05, which computes a 95% confidence interval.

...

Ignored.

Value

An data frame of state probabilities and data point estimates along with confidence intervals.

Examples

data(dax_model_3t)
predict(dax_model_3t)
#>   state_1 state_2 state_3       lb estimate      ub
#> 1 0.00515 0.97315 0.02170 -0.02190 -0.00023 0.02145
#> 2 0.01006 0.94769 0.04225 -0.02179 -0.00020 0.02138
#> 3 0.01477 0.92354 0.06170 -0.02168 -0.00018 0.02132
#> 4 0.01926 0.90063 0.08011 -0.02158 -0.00016 0.02126
#> 5 0.02356 0.87890 0.09754 -0.02148 -0.00014 0.02121