Setting and checking parameters for the fHMM package

This function sets and checks parameters for the fHMM package.

Usage

fHMM_parameters(
  controls,
  Gamma = NULL,
  mus = NULL,
  sigmas = NULL,
  dfs = NULL,
  Gammas_star = NULL,
  mus_star = NULL,
  sigmas_star = NULL,
  dfs_star = NULL,
  seed = NULL,
  scale_par = c(1, 1)
)

Arguments

controls

An object of class fHMM_controls.

Gamma

A tpm (transition probability matrix) of dimension controls$states[1].

mus

A vector of expectations of length controls$states[1].

sigmas

A vector of standard deviations of length controls$states[1].

dfs

A vector of degrees of freedom of length controls$states[1]. Only relevant if sdd is a t-distribution.

Gammas_star

A list of length controls$states[1] of tpm's. Each tpm must be of dimension controls$states[2].

mus_star

A list of length controls$states[1] of vectors of expectations. Each vector must be of length controls$states[2].

sigmas_star

A list of length controls$states[1] of vectors of standard deviations. Each vector must be of length controls$states[2].

dfs_star

A list of length controls$states[1] of vectors of degrees of freedom. Each vector must be of length controls$states[2]. Only relevant if sdd is a t-distribution.

seed

Set a seed for the sampling of parameters.

scale_par

A positive numeric vector of length two, containing scales for sampled expectations and standard deviations. The first entry is the scale for mus and sigmas, the second entry is the scale for mus_star and sigmas_star. Set an entry to 1 for no scaling.

Value

An object of class fHMM_parameters.

Details

See the vignettes for more information on how to specify parameters.

Examples

controls <- set_controls()
fHMM_parameters(controls)
#> fHMM parameters