Find closest year

This function takes a date as input and returns the closest year.

Usage

find_closest_year(date)

Arguments

date

[character(1)]
The date in format "YYYY-MM-DD".

Value

An integer, the closest year to the input date.