Simulate state-dependent observations

This function simulates state-dependent observations.

Usage

simulate_observations(
  markov_chain,
  sdd,
  mus,
  sigmas,
  dfs = NULL,
  seed = NULL,
  total_length = length(markov_chain)
)

Arguments

markov_chain

A numeric vector of states of a Markov chain.

sdd

The name of the state-dependent distribution, one of "t", "gamma", and "lnorm".

mus

A vector of expected values.

sigmas

A vector of standard deviations.

dfs

A vector of degrees of freedom (only relevant if sdd = "t").

seed

Set a seed.

total_length

An integer, the total length of the output vector. Must be greater or equal than length(markov_chain).

Value

A numeric vector of length total_length, where the first length(markov_chain) elements are numeric values and the last total_length - length(markov_chain) elements are NA_real_.